proposal create conditions are changed

+++ app/controllers/api/v1/trading_parcels_controller.rb
@@ -266,8 +266,9 @@ module Api
 
       def check_overdue_and_market_limit(transaction, parcel)
         buyer = Company.where(id: transaction.buyer_id).first
-        market_limit = get_market_limit_from_credit_limit_table(buyer, current_company).to_f
-        if market_limit < parcel.total_value.to_f || current_company.has_overdue_transaction_of_30_days(transaction.buyer_id)
+        market_limit = CreditLimit.where(seller_id: current_company.id, buyer_id: buyer.id).first
+        if (market_limit.present? && market_limit.market_limit.to_f < parcel.total_value.to_f) || current_company.has_overdue_transaction_of_30_days(transaction.buyer_id)
+          # if market_limit < parcel.total_value.to_f || current_company.has_overdue_transaction_of_30_days(transaction.buyer_id)
           parcel.destroy
           render json: { sucess: false, message: "This Company does not meet your risk parameters. Do you wish to cancel the transaction?" }
         else
